refactor: extract D&D 5e ability handlers from service
- Move hardcoded ability logic to dnd5e_handlers.go
- Remove handleRage, handleSecondWind, handleBardicInspiration,
  handleLayOnHands, handleDivineSense from service.go
- This is a temporary solution before implementing proper
  rulebook registration (see #242)
- Service is now less coupled to D&D 5e specific logic

Partially addresses #237
